requirements.txt
tensorboard
jupyterlab
ipywidgets
torchvision
pytorch_lightning
lightning
neptune
neptune-tensorboard
tqdm
